Update model.rs
wei is too big

+++ backend/src/db/model.rs
@@ -3,8 +3,8 @@
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct History {
     pub manager: String,
-    pub ticket_sold: i32,
-    pub game_money: i32,
+    pub ticket_sold: u64,
+    pub game_money: u64,
     pub lucky_number: String,
     // #[serde(skip_serializing_if = "Option::is_none")]
     pub first_winner: Vec<String>,
